Made servicelayer for @Postmapping in ParkingEventController.
