+++ src/main/java/com/johanekstroem/parking/Controller/ParkingEventController.java
@@ -4,18 +4,17 @@ import java.net.URI;
 import java.time.LocalDateTime;
 import java.util.List;
 
+
+import com.johanekstroem.parking.Service.ParkingEventService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import com.johanekstroem.parking.Entities.Car;
+
 import com.johanekstroem.parking.Entities.ParkingEvent;
-import com.johanekstroem.parking.Entities.ParkingSpot;
-import com.johanekstroem.parking.Repositories.CarRepository;
 import com.johanekstroem.parking.Repositories.ParkingEventRepository;
-import com.johanekstroem.parking.Repositories.ParkingSpotRepository;
 import com.johanekstroem.parking.Service.ValidateStopTime;
 
 @AllArgsConstructor
@@ -23,45 +22,26 @@ import com.johanekstroem.parking.Service.ValidateStopTime;
 @RequestMapping("/api")
 public class ParkingEventController {
 
+ValidateStopTime validateStopTime;
+ParkingEventService parkingEventService;
+ParkingEventRepository parkingEventRepository;
 
-  CarRepository carRepository;
-  ParkingSpotRepository parkingSpotRepository;
-  ParkingEventRepository parkingEventRepository;
-  ValidateStopTime validateStopTime;
   
 
   @PostMapping("/parkingevent")
   public Object startParking(@ModelAttribute("parkingevent") ParkingEvent parkingEvent) {
-    var stoptime = parkingEvent.getStoptime();
-    var saveParkingEvent = parkingEventRepository.save(parkingEvent);
+    if (ParkingEventService.isParkingtimeOk(parkingEvent)) {
+      ParkingEventService.saveParkingEvent(parkingEvent);
+      parkingEventRepository.save(parkingEvent);
 
-    if (validateStopTime.isInFuture(stoptime)) {
-      Long carID = parkingEvent.getCar().getId();
-      var carOptional = carRepository.findById(carID);
-      if (carOptional.isPresent()) {
-        Car car = carOptional.get();
-        car.addParkingEvent(parkingEvent);
-        carRepository.save(car);
-      }
-
-      Long parkingSpotID = parkingEvent.getParkingSpot().getId();
-      var parkingSpotOptional = parkingSpotRepository.findById(parkingSpotID);
-      if (parkingSpotOptional.isPresent()) {
-        ParkingSpot parkingSpot = parkingSpotOptional.get();
-        parkingSpot.addParkingEvent(parkingEvent);
-        parkingSpotRepository.save(parkingSpot);
-      
-         URI location = ServletUriComponentsBuilder
-                .fromCurrentRequest()
-                .buildAndExpand(saveParkingEvent.getId())
-                .toUri();
-                return "redirect:/saved";
-              }
+      return "redirect:/saved";
     }
 
     return "redirect:/ops";
   }
-  
+
+
+
   @GetMapping("/parkingevent")
   public Iterable<ParkingEvent> getAllParkingEvents() {
     return parkingEventRepository.findAll();
